system to generate and maintain customer accounts
this program can be used by a bank employee to create accounts for their customers. the customer details can be viewed and can be updated, and money can be deposited and withdrawn from their account. a list of all customers can be viewed at once, including only the important details. accounts can be deleted from the system as well. however, since this program does not use file handling, the date which is input, can only be accessed while the terminal is open.